Refactor: Please clean this
