+++ app/api/discord/events/route.js
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server"
-import { verifyDiscordRequest } from "../../../../lib/discord"
-import { startVoiceSession, endVoiceSession, recordMediaSubmission } from "../../../../lib/xp"
+import { verifyDiscordRequest } from "../../../../lib/discord/verifyRequest"
+import { startVoiceSession, endVoiceSession, recordMediaSubmission } from "../../../../lib/discord/commands" // xp.js and commands.js are the exact same file; how does this import statement even compile??? you should use TS because this entire project is all over the place
 
 export async function POST(req) {
   try {
