+++ lib/db/prismaClient.js
@@ -1,0 +1,3 @@
+import { PrismaClient } from "@prisma/client"
+const prisma = new PrismaClient()
+export default prisma
