+++ lib/discord/commands.js
