+++ lib/discord/points.js
@@ -1,5 +1,5 @@
 import prisma from "./db"
-import { discordRequest } from "./discord"
+import { discordRequest } from "./verifyRequest"
 
 // Get points for a user in a guild
 export async function getUserPoints(guildId, userId) {
