+++ lib/discord/starboard.js
@@ -1,5 +1,5 @@
-import { discordRequest } from "./discord"
-import prisma from "./db"
+import { discordRequest } from "./verifyRequest"
+import prisma from "../db/prismaClient"
 
 // Set up starboard configuration for a guild
 export async function setStarboardConfig(guildId, channelId, threshold = 3, emoji = "⭐") {
